rename(user) : 디렉토리 정리
