+++ src/main/java/com/logilink/auth/common/constants/UserStatus.java
@@ -1,4 +1,4 @@
-package com.logilink.auth.model.entity;
+package com.logilink.auth.common.constants;
 
 public enum UserStatus {
