+++ src/main/java/com/logilink/auth/model/dto/request/UserSignupReq.java
@@ -1,7 +1,7 @@
 package com.logilink.auth.model.dto.request;
 
 import com.logilink.auth.model.dto.UserSignupInfo;
-import com.logilink.auth.model.entity.UserRole;
+import com.logilink.auth.common.constants.UserRole;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
