+++ src/main/java/com/logilink/auth/model/dto/response/UserSignupRes.java
@@ -1,8 +1,6 @@
 package com.logilink.auth.model.dto.response;
 
 import com.logilink.auth.model.entity.User;
-import com.logilink.auth.model.entity.UserRole;
-import com.logilink.auth.model.entity.UserStatus;
 
 public record UserSignupRes(
     String username,
