+++ src/main/java/com/logilink/auth/service/UserService.java
@@ -1,17 +1,16 @@
 package com.logilink.auth.service;
 
-import static com.logilink.auth.exception.UserErrorCode.*;
-import static com.logilink.auth.model.entity.UserStatus.*;
+import static com.logilink.auth.common.constants.UserStatus.*;
+import static com.logilink.auth.common.exception.UserErrorCode.*;
 
 import com.logilink.auth.auth.JwtUtil;
 import com.logilink.auth.config.security.LoginMasterKeyConfig;
-import com.logilink.auth.exception.AppException;
+import com.logilink.auth.common.exception.AppException;
 import com.logilink.auth.model.dto.UserInfo;
 import com.logilink.auth.model.dto.UserMyInfo;
 import com.logilink.auth.model.dto.UserSignupInfo;
 import com.logilink.auth.model.dto.request.MasterSignupReq;
 import com.logilink.auth.model.dto.request.UserLoginReq;
-import com.logilink.auth.model.dto.request.UserSearchReq;
 import com.logilink.auth.model.dto.request.UserSignupReq;
 import com.logilink.auth.model.dto.request.UserStatusUpdateReq;
 import com.logilink.auth.model.dto.request.UserUpdateReq;
@@ -20,19 +19,17 @@ import com.logilink.auth.model.dto.response.UserLoginRes;
 import com.logilink.auth.model.dto.response.UserPageRes;
 import com.logilink.auth.model.dto.response.UserSignupRes;
 import com.logilink.auth.model.dto.response.UserStatusUpdateRes;
-import com.logilink.auth.model.entity.DeliveryType;
+import com.logilink.auth.common.constants.DeliveryType;
 import com.logilink.auth.model.entity.DeliveryUser;
 import com.logilink.auth.model.entity.User;
-import com.logilink.auth.model.entity.UserRole;
-import com.logilink.auth.model.entity.UserStatus;
+import com.logilink.auth.common.constants.UserRole;
+import com.logilink.auth.common.constants.UserStatus;
 import com.logilink.auth.repository.DeliveryUserRepository;
 import com.logilink.auth.repository.UserRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -65,7 +62,7 @@ public class UserService {
     public MasterSignupRes masterSignup(String secretKey, MasterSignupReq masterSignupReq) {
         // 헤더에서 가져온 시크릿 키 검증
         if (!secretKey.equals(loginMasterKeyConfig.getMasterSecretKey())) {
-            throw new AppException(INVALID_SECRET_KEY);
+            throw AppException.of(INVALID_SECRET_KEY);
         }
         // 중복 데이터 검증
         checkDuplicateUser(masterSignupReq);
@@ -84,11 +81,11 @@ public class UserService {
     public UserLoginRes login(UserLoginReq loginReq) {
         // 유저 확인
         User user = userRepository.findValidUserByUsername(loginReq.username())
-            .orElseThrow(() -> new AppException(INVALID_LOGIN));
+            .orElseThrow(() -> AppException.of(INVALID_LOGIN));
 
         // 비밀번호 검증
         if (!passwordEncoder.matches(loginReq.password(), user.getPassword())) {
-            throw new AppException(INVALID_LOGIN);
+            throw AppException.of(INVALID_LOGIN);
         }
 
         // 승인 상태 확인
@@ -102,8 +99,9 @@ public class UserService {
     }
 
     @Transactional
-    public UserStatusUpdateRes updateUserStatusByMaster(User user,
-        UserStatusUpdateReq statusUpdateReq) {
+    public UserStatusUpdateRes updateUserStatusByMaster(
+        User user, UserStatusUpdateReq statusUpdateReq
+    ) {
         // 권한 확인
         checkMasterRole(user);
 
@@ -121,8 +119,9 @@ public class UserService {
     }
 
     @Transactional
-    public UserStatusUpdateRes updateUserStatusByHubManager(User user,
-        UserStatusUpdateReq statusUpdateReq) {
+    public UserStatusUpdateRes updateUserStatusByHubManager(
+        User user, UserStatusUpdateReq statusUpdateReq
+    ) {
         // 권한 확인
         checkHubManagerRole(user);
 
@@ -231,9 +230,10 @@ public class UserService {
         return UserMyInfo.from(me);
     }
 
+
     private User getUser(Long userId) {
         return userRepository.findValidUserById(userId)
-            .orElseThrow(() -> new AppException(USER_NOT_FOUND));
+            .orElseThrow(() -> AppException.of(USER_NOT_FOUND));
     }
 
     private User generateUser(UserSignupReq signupReq, String encodedPassword, UserStatus status) {
@@ -253,29 +253,29 @@ public class UserService {
 
     private static void checkMasterRole(User user) {
         if (user.getRole() != UserRole.MASTER) {
-            throw new AppException(REQUIRE_MASTER_ROLE);
+            throw AppException.of(REQUIRE_MASTER_ROLE);
         }
     }
 
     private static void checkHubManagerRole(User user) {
         if (user.getRole() != UserRole.HUB_MANAGER) {
-            throw new AppException(REQUIRE_HUB_MASTER_ROLE);
+            throw AppException.of(REQUIRE_HUB_MASTER_ROLE);
         }
     }
 
     private void checkDuplicateUser(UserSignupInfo signupInfo) {
         if (userRepository.existsValidUserByUsername(signupInfo.username())) {
-            throw new AppException(DUPLICATE_USERNAME);
+            throw AppException.of(DUPLICATE_USERNAME);
         }
 
         if (userRepository.existsValidUserByEmail(signupInfo.email())) {
-            throw new AppException(DUPLICATE_EMAIL);
+            throw AppException.of(DUPLICATE_EMAIL);
         }
     }
 
     private void checkUserStatus(User user) {
         if (user.getUserStatus() != UserStatus.APPROVED) {
-            throw new AppException(USER_NOT_APPROVED);
+            throw AppException.of(USER_NOT_APPROVED);
         }
     }
 }
